+++ src/Core/ZakYip.WheelDiverterSorter.Core/Sorting/Contracts/AssignChuteContracts.cs
@@ -58,7 +58,7 @@ public record AssignChuteResponse
     /// <summary>
     /// 目标格口ID
     /// </summary>
-    public required int ChuteId { get; init; }
+    public required long ChuteId { get; init; }
 
     /// <summary>
     /// 是否成功
